HTML and CSS Basics Focus on Email Development
In this repository, you can find the presentation in PDF format that I made when I was doing training for the DEPT employees for HTML and CSS basics with a focus on Email development for CRM platforms. Also, two ready-to-use and test Email Template examples that can be adjusted for your needs.